// packages/ts/src/types/ast.ts
export interface ArrayLiteralExpression extends ts.Node {
	readonly elements: ts.NodeArray<Expression>;
	readonly kind: ts.SyntaxKind.ArrayLiteralExpression;
	readonly parent: LeftHandSideExpressionParent;
}
export interface ArrayTypeNode extends ts.Node {
	readonly elementType: TypeNode;
	readonly kind: ts.SyntaxKind.ArrayType;
	readonly parent: TypeNodeParent;
}
export interface ArrowFunction extends ts.Node {
	readonly asteriskToken?: AsteriskToken | undefined;
	readonly body: ConciseBody;
	readonly equalsGreaterThanToken: EqualsGreaterThanToken;
	readonly exclamationToken?: ExclamationToken | undefined;
	readonly kind: ts.SyntaxKind.ArrowFunction;
	readonly modifiers?: ts.NodeArray<Modifier>;
	readonly name: never;
	readonly parameters: ts.NodeArray<ParameterDeclaration>;
	readonly parent: ExpressionParent;
	readonly questionToken?: QuestionToken | undefined;
	readonly type?: TypeNode | undefined;
	readonly typeParameters?: ts.NodeArray<TypeParameterDeclaration> | undefined;
}
export interface AsExpression extends ts.Node {
	readonly expression: Expression;
	readonly kind: ts.SyntaxKind.AsExpression;
	readonly parent: ExpressionParent;
	readonly type: TypeNode;
}
// ...
export interface AwaitExpression extends ts.Node {
	readonly expression: UnaryExpression;
	readonly kind: ts.SyntaxKind.AwaitExpression;
	readonly parent:
		| AwaitExpression
		| DeleteExpression
		| ExpressionParent
		| PrefixUnaryExpression
		| TypeAssertion
		| TypeOfExpression
		| VoidExpression;
}
// ...
export interface BinaryExpression extends ts.Node {
	readonly kind: ts.SyntaxKind.BinaryExpression;
	readonly left: Expression;
	readonly operatorToken: BinaryOperatorToken;
	readonly parent: ExpressionParent;
	readonly right: Expression;
}
export type BinaryOperator =
	| AssignmentOperatorOrHigher
	| ts.SyntaxKind.CommaToken;
export interface BinaryOperatorToken extends ts.Node {
	readonly kind: BinaryOperator;
	readonly parent: BinaryExpression;
}
export interface BindingElement extends ts.Node {
	readonly dotDotDotToken?: DotDotDotToken;
	readonly initializer?: Expression;
	readonly kind: ts.SyntaxKind.BindingElement;
	readonly name: BindingName;
	readonly parent: BindingPattern;
	readonly propertyName?: PropertyName;
}
// ...
export interface BreakStatement extends ts.Node {
	readonly kind: ts.SyntaxKind.BreakStatement;
	readonly label?: Identifier;
	readonly parent: StatementParent;
}
export interface CallExpression extends ts.Node {
	readonly arguments: ts.NodeArray<Expression>;
	readonly expression: LeftHandSideExpression;
	readonly kind: ts.SyntaxKind.CallExpression;
	readonly parent: LeftHandSideExpressionParent;
	readonly questionDotToken?: QuestionDotToken;
	readonly typeArguments?: ts.NodeArray<TypeNode>;
}
// ...
export interface CaseClause extends ts.Node {
	readonly expression: Expression;
	readonly kind: ts.SyntaxKind.CaseClause;
	readonly parent: CaseBlock;
	readonly statements: ts.NodeArray<Statement>;
}
// ...
export interface ClassExpression extends ts.Node {
	readonly heritageClauses?: ts.NodeArray<HeritageClause>;
	readonly kind: ts.SyntaxKind.ClassExpression;
	readonly members: ts.NodeArray<ClassElement>;
	readonly modifiers?: ts.NodeArray<ModifierLike>;
	readonly name?: Identifier;
	readonly parent: LeftHandSideExpressionParent;
	readonly typeParameters?: ts.NodeArray<TypeParameterDeclaration>;
}
export interface ClassStaticBlockDeclaration extends ts.Node {
	readonly body: Block;
	readonly kind: ts.SyntaxKind.ClassStaticBlockDeclaration;
	readonly name?: PropertyName;
	readonly parent: ClassDeclaration | ClassExpression;
}
export interface ColonToken extends ts.Node {
	readonly kind: ts.SyntaxKind.ColonToken;
	readonly parent: ConditionalExpression;
}
export interface CommaListExpression extends ts.Node {
	readonly elements: ts.NodeArray<Expression>;
	readonly kind: ts.SyntaxKind.CommaListExpression;
	readonly parent: ExpressionParent;
}
// ...
export interface ComputedPropertyName extends ts.Node {
	readonly expression: Expression;
	readonly kind: ts.SyntaxKind.ComputedPropertyName;
	readonly parent:
		| BindingElement
		| ClassStaticBlockDeclaration
		| ConstructorDeclaration
		| EnumMember
		| FunctionOrConstructorTypeNodeBase
		| JSDocFunctionType
		| JsxSpreadAttribute
		| MethodDeclaration
		| PropertyAssignment
		| PropertyDeclaration
		| SemicolonClassElement
		| SpreadAssignment
		| TypeElement;
}
export type ConciseBody = Expression | FunctionBody;
export interface ConditionalExpression extends ts.Node {
	readonly colonToken: ColonToken;
	readonly condition: Expression;
	readonly kind: ts.SyntaxKind.ConditionalExpression;
	readonly parent: ExpressionParent;
	readonly questionToken: QuestionToken;
	readonly whenFalse: Expression;
	readonly whenTrue: Expression;
}
export interface ConditionalTypeNode extends ts.Node {
	readonly checkType: TypeNode;
	readonly extendsType: TypeNode;
	readonly falseType: TypeNode;
	readonly kind: ts.SyntaxKind.ConditionalType;
	readonly parent: TypeNodeParent;
	readonly trueType: TypeNode;
}
export interface ConstKeyword extends ts.Node {
	readonly kind: ts.SyntaxKind.ConstKeyword;
	readonly parent: ModifierParent;
}
export interface ConstructorDeclaration extends ts.Node {
	readonly asteriskToken?: AsteriskToken | undefined;
	readonly body?: FunctionBody | undefined;
	readonly exclamationToken?: ExclamationToken | undefined;
	readonly kind: ts.SyntaxKind.Constructor;
	readonly modifiers?: ts.NodeArray<ModifierLike> | undefined;
	readonly name?: PropertyName;
	readonly parameters: ts.NodeArray<ParameterDeclaration>;
	readonly parent: ClassDeclaration | ClassExpression;
	readonly questionToken?: QuestionToken | undefined;
	readonly type?: TypeNode | undefined;
	readonly typeParameters?: ts.NodeArray<TypeParameterDeclaration> | undefined;
}
// ...
export interface ContinueStatement extends ts.Node {
	readonly kind: ts.SyntaxKind.ContinueStatement;
	readonly label?: Identifier;
	readonly parent: StatementParent;
}
export interface DebuggerStatement extends ts.Node {
	readonly kind: ts.SyntaxKind.DebuggerStatement;
	readonly parent: StatementParent;
}
export type DeclarationStatement =
	| ClassDeclaration
	| EnumDeclaration
	| ExportAssignment
	| ExportDeclaration
	| FunctionDeclaration
	| ImportEqualsDeclaration
	| InterfaceDeclaration
	| MissingDeclaration
	| ModuleDeclaration
	| NamespaceExportDeclaration
	| TypeAliasDeclaration;
export interface DeclareKeyword extends ts.Node {
	readonly kind: ts.SyntaxKind.DeclareKeyword;
	readonly parent: ModifierParent;
}
export interface Decorator extends ts.Node {
	readonly expression: LeftHandSideExpression;
	readonly kind: ts.SyntaxKind.Decorator;
	readonly parent:
		| AccessorDeclaration
		| ClassDeclaration
		| ClassExpression
		| ConstructorDeclaration
		| EnumDeclaration
		| ExportAssignment
		| ExportDeclaration
		| FunctionDeclaration
		| ImportDeclaration
		| ImportEqualsDeclaration
		| IndexSignatureDeclaration
		| InterfaceDeclaration
		| JSDocNamespaceDeclaration
		| MethodDeclaration
		| ModuleDeclaration
		| NamespaceDeclaration
		| ParameterDeclaration
		| PropertyDeclaration
		| TypeAliasDeclaration
		| VariableStatement;
}
export interface DefaultClause extends ts.Node {
	readonly kind: ts.SyntaxKind.DefaultClause;
	readonly parent: CaseBlock;
	readonly statements: ts.NodeArray<Statement>;
}
export interface DefaultKeyword extends ts.Node {
	readonly kind: ts.SyntaxKind.DefaultKeyword;
	readonly parent: ModifierParent;
}
export interface DeleteExpression extends ts.Node {
	readonly expression: UnaryExpression;
	readonly kind: ts.SyntaxKind.DeleteExpression;
	readonly parent:
		| AwaitExpression
		| DeleteExpression
		| ExpressionParent
		| PrefixUnaryExpression
		| TypeAssertion
		| TypeOfExpression
		| VoidExpression;
}
export interface DoStatement extends ts.Node {
	readonly expression: Expression;
	readonly kind: ts.SyntaxKind.DoStatement;
	readonly parent: StatementParent;
	readonly statement: Statement;
}
export interface DotDotDotToken extends ts.Node {
	readonly kind: ts.SyntaxKind.DotDotDotToken;
	readonly parent: BindingElement | ParameterDeclaration;
}
export interface ElementAccessExpression extends ts.Node {
	readonly argumentExpression: Expression;
	readonly expression: LeftHandSideExpression;
	readonly kind: ts.SyntaxKind.ElementAccessExpression;
	readonly parent: LeftHandSideExpressionParent;
	readonly questionDotToken?: QuestionDotToken;
}
// ...
export interface EnumMember extends ts.Node {
	readonly initializer?: Expression;
	readonly kind: ts.SyntaxKind.EnumMember;
	readonly name: PropertyName;
	readonly parent: EnumDeclaration;
}
// ...
export interface ExportSpecifier extends ts.Node {
	readonly isTypeOnly: boolean;
	readonly kind: ts.SyntaxKind.ExportSpecifier;
	readonly name: ModuleExportName;
	readonly parent: NamedExports;
	readonly propertyName?: ModuleExportName;
}
export type Expression =
	| ArrowFunction
	| AsExpression
	| BinaryExpression
	| CommaListExpression
	| ConditionalExpression
	| JsxClosingFragment
	| JsxExpression
	| JsxOpeningElement
	| JsxOpeningFragment
	| OmittedExpression
	| SatisfiesExpression
	| SpreadElement
	| SyntheticExpression
	| UnaryExpression
	| YieldExpression;
export interface ExpressionStatement extends ts.Node {
	readonly expression: Expression;
	readonly kind: ts.SyntaxKind.ExpressionStatement;
	readonly parent: StatementParent;
}
// ...
export interface ExternalModuleReference extends ts.Node {
	readonly expression: Expression;
	readonly kind: ts.SyntaxKind.ExternalModuleReference;
	readonly parent: ImportEqualsDeclaration;
}
// ...
export type FunctionBody = Block;
// ...
export interface Identifier extends ts.Node {
	readonly kind: ts.SyntaxKind.Identifier;
	readonly parent:
		| ArrayLiteralExpression
		| ArrowFunction
		| AsExpression
		| AwaitExpression
		| BinaryExpression
		| BindingElement
		| BreakStatement
		| CallExpression
		| CaseClause
		| ClassExpression
		| ClassStaticBlockDeclaration
		| CommaListExpression
		| ComputedPropertyName
		| ConditionalExpression
		| ConstructorDeclaration
		| ContinueStatement
		| DeclarationStatement
		| Decorator
		| DeleteExpression
		| ElementAccessExpression
		| EnumMember
		| ExportSpecifier
		| ExpressionStatement
		| ExternalModuleReference
		| FunctionExpression
		| FunctionOrConstructorTypeNodeBase
		| IfStatement
		| ImportAttribute
		| ImportClause
		| ImportDeclaration
		| ImportSpecifier
		| IterationStatement
		| JSDocFunctionType
		| JSDocLink
		| JSDocLinkCode
		| JSDocLinkPlain
		| JSDocMemberName
		| JSDocNamespaceDeclaration
		| JSDocPropertyLikeTag
		| JSDocReturnTag
		| JSDocTemplateTag
		| JsxAttributeLike
		| JsxClosingElement
		| JsxExpression
		| JsxNamespacedName
		| JsxOpeningElement
		| JsxSelfClosingElement
		| JsxTagNamePropertyAccess
		| LabeledStatement
		| MetaProperty
		| MethodDeclaration
		| NamedTupleMember
		| NamespaceDeclaration
		| NamespaceExport
		| NamespaceImport
		| NewExpression
		| NodeWithTypeArguments
		| NonNullExpression
		| ParameterDeclaration
		| ParenthesizedExpression
		| PartiallyEmittedExpression
		| PostfixUnaryExpression
		| PrefixUnaryExpression
		| PropertyAccessExpression
		| PropertyAssignment
		| PropertyDeclaration
		| QualifiedName
		| ReturnStatement
		| SatisfiesExpression
		| SemicolonClassElement
		| ShorthandPropertyAssignment
		| SpreadAssignment
		| SpreadElement
		| SwitchStatement
		| TaggedTemplateExpression
		| TemplateSpan
		| ThrowStatement
		| TypeAssertion
		| TypeElement
		| TypeOfExpression
		| TypeParameterDeclaration
		| TypePredicateNode
		| VariableDeclaration
		| VoidExpression
		| WithStatement
		| YieldExpression;

	/**
	 * Prefer to use `id.unescapedText`. (Note: This is available only in services, not internally to the TypeScript compiler.)
	 * Text of identifier, but if the identifier begins with two underscores, this will begin with three.
	 */
	readonly escapedText: ts.__String;
	readonly text: string;
}
// ...
export interface SourceFile extends ts.Node {
	readonly kind: ts.SyntaxKind.SourceFile;
	// parent is actually undefined, see comment for NullNode
	amdDependencies: readonly ts.AmdDependency[];
	readonly endOfFileToken: Token<ts.SyntaxKind.EndOfFileToken, SourceFile>;
	fileName: string;
	isDeclarationFile: boolean;
	languageVariant: ts.LanguageVariant;
	libReferenceDirectives: readonly ts.FileReference[];
	moduleName?: string;
	readonly parent: NullNode;
	referencedFiles: readonly ts.FileReference[];
	readonly statements: ts.NodeArray<Statement>;
	text: string;
	typeReferenceDirectives: readonly ts.FileReference[];

	/**
	 * lib.d.ts should have a reference comment like
	 *
	 *  /// &lt;reference no-default-lib="true"/>
	 *
	 * If any other file has this comment, it signals not to include lib.d.ts
	 * because this containing file is intended to act as a default library.
	 */
	hasNoDefaultLib: boolean;
	languageVersion: ts.ScriptTarget;

	/**
	 * When `module` is `Node16` or `NodeNext`, this field controls whether the
	 * source file in question is an ESNext-output-format file, or a CommonJS-output-format
	 * module. This is derived by the module resolver as it looks up the file, since
	 * it is derived from either the file extension of the module, or the containing
	 * `package.json` context, and affects both checking and emit.
	 *
	 * It is _public_ so that (pre)transformers can set this field,
	 * since it switches the builtin `node` module transform. Generally speaking, if unset,
	 * the field is treated as though it is `ModuleKind.CommonJS`.
	 *
	 * Note that this field is only set by the module resolution process when
	 * `moduleResolution` is `Node16` or `NodeNext`, which is implied by the `module` setting
	 * of `Node16` or `NodeNext`, respectively, but may be overriden (eg, by a `moduleResolution`
	 * of `node`). If so, this field will be unset and source files will be considered to be
	 * CommonJS-output-format by the node module transformer and type checker, regardless of extension or context.
	 */
	getLineAndCharacterOfPosition(pos: number): ts.LineAndCharacter;
	getLineEndOfPosition(pos: number): number;
	getLineStarts(): readonly number[];
	getPositionOfLineAndCharacter(line: number, character: number): number;
	impliedNodeFormat?: ts.ResolutionMode;
	update(newText: string, textChangeRange: ts.TextChangeRange): SourceFile;
}